Hide StorageClient from public API surface
- move newPipeline() out to an export function for File and Queue

- update tests, samples, and readme
