+++ sdk/storage/storage-blob/samples/javascript/basic.js
@@ -4,7 +4,7 @@
 
 const {
   BlobServiceClient,
-  StorageClient,
+  newPipeline,
   SharedKeyCredential,
   AnonymousCredential,
   TokenCredential
