+++ sdk/storage/storage-blob/samples/javascript/iterator.js
@@ -2,7 +2,7 @@
  Setup: Enter your storage account name and shared key in main()
 */
 
-const { BlobServiceClient, StorageClient, SharedKeyCredential } = require("../.."); // Change to "@azure/storage-blob" in your package
+const { BlobServiceClient, newPipeline, SharedKeyCredential } = require("../.."); // Change to "@azure/storage-blob" in your package
 
 async function main() {
   // Enter your storage account name and shared key
