+++ sdk/storage/storage-blob/samples/typescript/advanced.ts
@@ -5,6 +5,7 @@
 import fs from "fs";
 import {
   Aborter,
+  AnonymousCredential,
   BlobServiceClient,
   newPipeline
 } from "../../src"; // Change to "@azure/storage-blob" in your package
