+++ sdk/storage/storage-blob/src/StorageClient.ts
@@ -34,7 +34,7 @@ export abstract class StorageClient {
    * generated by AutoRest generator.
    *
    * @protected
-   * @type {StorageClient}
+   * @type {StorageClientContext}
    * @memberof StorageClient
    */
   protected readonly storageClientContext: StorageClientContext;
