+++ sdk/storage/storage-file/samples/javascript/advanced.js
@@ -7,7 +7,7 @@ const {
   AnonymousCredential,
   Aborter,
   FileServiceClient,
-  StorageClient
+  newPipeline
 } = require("../.."); // Change to "@azure/storage-file" in your package
 
 async function main() {
@@ -16,7 +16,7 @@ async function main() {
   const accountSas = "";
   const localFilePath = "";
 
-  const pipeline = StorageClient.newPipeline(new AnonymousCredential(), {
+  const pipeline = newPipeline(new AnonymousCredential(), {
     // httpClient: MyHTTPClient, // A customized HTTP client implementing IHttpClient interface
     // logger: MyLogger, // A customized logger implementing IHttpPipelineLogger interface
     retryOptions: { maxTries: 4 }, // Retry options
