+++ sdk/storage/storage-file/samples/javascript/basic.js
@@ -3,7 +3,7 @@
 */
 
 const {
-  StorageClient,
+  newPipeline,
   FileServiceClient,
   SharedKeyCredential,
   AnonymousCredential
@@ -22,7 +22,7 @@ async function main() {
   const anonymousCredential = new AnonymousCredential();
 
   // Use sharedKeyCredential or anonymousCredential to create a pipeline
-  const pipeline = StorageClient.newPipeline(sharedKeyCredential);
+  const pipeline = newPipeline(sharedKeyCredential);
 
   // List shares
   const serviceClient = new FileServiceClient(
