+++ sdk/storage/storage-file/src/StorageClient.ts
@@ -1,57 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import { deserializationPolicy, RequestPolicyFactory } from "@azure/ms-rest-js";
-
-import { BrowserPolicyFactory } from "./BrowserPolicyFactory";
-import { Credential } from "./credentials/Credential";
 import { StorageClientContext } from "./generated/lib/storageClientContext";
-import { LoggingPolicyFactory } from "./LoggingPolicyFactory";
-import { IHttpClient, IHttpPipelineLogger, Pipeline } from "./Pipeline";
-import { RetryOptions, RetryPolicyFactory } from "./RetryPolicyFactory";
-import { TelemetryOptions, TelemetryPolicyFactory } from "./TelemetryPolicyFactory";
-import { UniqueRequestIDPolicyFactory } from "./UniqueRequestIDPolicyFactory";
+import { Pipeline } from "./Pipeline";
 import { escapeURLPath } from "./utils/utils.common";
-
-export { deserializationPolicy };
-
-/**
- * Option interface for Pipeline.newPipeline method.
- *
- * @export
- * @interface NewPipelineOptions
- */
-export interface NewPipelineOptions {
-  /**
-   * Telemetry configures the built-in telemetry policy behavior.
-   *
-   * @type {TelemetryOptions}
-   * @memberof NewPipelineOptions
-   */
-  telemetry?: TelemetryOptions;
-  /**
-   * Configures the built-in retry policy behavior.
-   *
-   * @type {RetryOptions}
-   * @memberof NewPipelineOptions
-   */
-  retryOptions?: RetryOptions;
-
-  /**
-   * Configures the HTTP pipeline logger.
-   *
-   * @type {IHttpPipelineLogger}
-   * @memberof NewPipelineOptions
-   */
-  logger?: IHttpPipelineLogger;
-  /**
-   * Configures the HTTP client to send requests and receive responses.
-   *
-   * @type {IHttpClient}
-   * @memberof NewPipelineOptions
-   */
-  httpClient?: IHttpClient;
-}
 
 /**
  * A ServiceClient represents a based URL class for ServiceClient, ContainerURL and etc.
@@ -60,38 +12,6 @@ export interface NewPipelineOptions {
  * @class StorageClient
  */
 export abstract class StorageClient {
-  /**
-   * A static method used to create a new Pipeline object with Credential provided.
-   *
-   * @static
-   * @param {Credential} credential Such as AnonymousCredential, SharedKeyCredential.
-   * @param {NewPipelineOptions} [pipelineOptions] Optional. Options.
-   * @returns {Pipeline} A new Pipeline object.
-   * @memberof Pipeline
-   */
-  public static newPipeline(
-    credential: Credential,
-    pipelineOptions: NewPipelineOptions = {}
-  ): Pipeline {
-    // Order is important. Closer to the API at the top & closer to the network at the bottom.
-    // The credential's policy factory must appear close to the wire so it can sign any
-    // changes made by other factories (like UniqueRequestIDPolicyFactory)
-    const factories: RequestPolicyFactory[] = [
-      new TelemetryPolicyFactory(pipelineOptions.telemetry),
-      new UniqueRequestIDPolicyFactory(),
-      new BrowserPolicyFactory(),
-      deserializationPolicy(), // Default deserializationPolicy is provided by protocol layer
-      new RetryPolicyFactory(pipelineOptions.retryOptions),
-      new LoggingPolicyFactory(),
-      credential
-    ];
-
-    return new Pipeline(factories, {
-      HTTPClient: pipelineOptions.httpClient,
-      logger: pipelineOptions.logger
-    });
-  }
-
   /**
    * Request policy pipeline.
    *
@@ -114,7 +34,7 @@ export abstract class StorageClient {
    * generated by AutoRest generator.
    *
    * @protected
-   * @type {StorageClient}
+   * @type {StorageClientContext}
    * @memberof StorageClient
    */
   protected readonly storageClientContext: StorageClientContext;
