+++ sdk/storage/storage-file/src/index.browser.ts
@@ -22,5 +22,4 @@ export * from "./TelemetryPolicyFactory";
 export * from "./UniqueRequestIDPolicyFactory";
 export * from "./BrowserPolicyFactory";
 export * from "./FileServiceClient";
-export * from "./StorageClient";
 export { Models, RestError };
