+++ sdk/storage/storage-file/test/node/sas.spec.ts
@@ -5,17 +5,17 @@ import {
   AccountSASResourceTypes,
   AccountSASServices,
   AnonymousCredential,
-  generateAccountSASQueryParameters,
   FileServiceClient,
-  SharedKeyCredential,
-  StorageClient,
-  SASProtocol
+  generateAccountSASQueryParameters,
+  SASProtocol,
+  SharedKeyCredential
 } from "../../src";
-import { FileSASPermissions } from "../../src/FileSASPermissions";
 import { FileClient } from "../../src/FileClient";
+import { FileSASPermissions } from "../../src/FileSASPermissions";
 import { generateFileSASQueryParameters } from "../../src/FileSASSignatureValues";
-import { ShareSASPermissions } from "../../src/ShareSASPermissions";
+import { newPipeline } from "../../src/Pipeline";
 import { ShareClient } from "../../src/ShareClient";
+import { ShareSASPermissions } from "../../src/ShareSASPermissions";
 import { getBSU, getUniqueName } from "../utils";
 
 describe("Shared Access Signature (SAS) generation Node.js only", () => {
@@ -49,7 +49,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${serviceClient.url}?${sas}`;
     const serviceClientWithSAS = new FileServiceClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     await serviceClientWithSAS.listSharesSegment();
@@ -76,7 +76,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${serviceClient.url}?${sas}`;
     const serviceClientWithSAS = new FileServiceClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     let error;
@@ -110,7 +110,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${serviceClient.url}?${sas}`;
     const serviceClientWithSAS = new FileServiceClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     let error;
@@ -147,7 +147,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${serviceClient.url}?${sas}`;
     const serviceClientWithSAS = new FileServiceClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     let error;
@@ -191,7 +191,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${shareClient.url}?${shareSAS}`;
     const shareClientwithSAS = new ShareClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     const dirURLwithSAS = shareClientwithSAS.createDirectoryClient("");
@@ -249,7 +249,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${fileClient.url}?${fileSAS}`;
     const fileClientwithSAS = new FileClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     const properties = await fileClientwithSAS.getProperties();
@@ -312,7 +312,7 @@ describe("Shared Access Signature (SAS) generation Node.js only", () => {
     const sasURL = `${shareClient.url}?${shareSAS}`;
     const shareClientwithSAS = new ShareClient(
       sasURL,
-      StorageClient.newPipeline(new AnonymousCredential())
+      newPipeline(new AnonymousCredential())
     );
 
     const dirURLwithSAS = shareClientwithSAS.createDirectoryClient("");
