+++ sdk/storage/storage-queue/src/index.ts
@@ -30,6 +30,5 @@ export * from "./QueueSASPermissions";
 export * from "./QueueSASSignatureValues";
 export * from "./UniqueRequestIDPolicyFactory";
 export * from "./QueueServiceClient";
-export * from "./StorageClient";
 export * from "./SASQueryParameters";
 export { Models, RestError };
